requirements.txt
# ----------------------------
# Core Backend
# ----------------------------
fastapi==0.110.0
uvicorn[standard]==0.29.0

# ----------------------------
# File Uploads & Forms
# ----------------------------
python-multipart==0.0.9

# ----------------------------
# Environment Variables
# ----------------------------
python-dotenv==1.0.1

# ----------------------------
# OpenAI Vision API
# ----------------------------
openai==1.12.0

# ----------------------------
# HTTP & Networking
# ----------------------------
httpx==0.27.0
requests==2.31.0

# ----------------------------
# Security & Validation
# ----------------------------
pydantic==2.6.4
pydantic-settings==2.2.1

# ----------------------------
# Logging & Utilities
# ----------------------------
loguru==0.7.2

# ----------------------------
# Image Handling (Future-proofing)
# ----------------------------
pillow

# ----------------------------
# Optional (Production Hardening)
# ----------------------------
python-jose==3.3.0        # JWT auth (future)
slowapi==0.1.9            # Rate limiting
